pyproject.toml
[project]
name = "mlfts"
version = "0.1.0"
description = "MLflow with Databricks and Unity Catalog Trace Storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlflow @ git+https://github.com/smoorjani/mlflow.git@issue-discovery",
    "databricks-sdk>=0.20.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "litellm>=1.81.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]
